+++ internal/manager/run.go
@@ -59,7 +59,7 @@ func Run(ctx context.Context, c *Config, diagnostic util.ConfigDumpDiagnostic, d
 	if c.KongAdminToken != "" {
 		c.KongAdminAPIConfig.Headers = append(c.KongAdminAPIConfig.Headers, "kong-admin-token:"+c.KongAdminToken)
 	}
-	kongClients, err := getKongClients(ctx, c)
+	kongClients, err := getKongClients(ctx, c, setupLog)
 	if err != nil {
 		return fmt.Errorf("unable to build kong api client(s): %w", err)
 	}
@@ -95,7 +95,7 @@ func Run(ctx context.Context, c *Config, diagnostic util.ConfigDumpDiagnostic, d
 	}
 
 	setupLog.Info("Starting Admission Server")
-	if err := setupAdmissionServer(ctx, c, mgr.GetClient(), deprecatedLogger); err != nil {
+	if err := setupAdmissionServer(ctx, c, mgr.GetClient(), setupLog, deprecatedLogger); err != nil {
 		return err
 	}